+++ docs/Jupyter_Book/changelog.rst
@@ -8,9 +8,9 @@ V 0.6.2 (MAY-2024)
 
 Improvements
 ^^^^^^^^^^^^
-* New *clip* parameter in *to_dose* method to limit the maximum dose, (usefull for brachytherapy film dosimetry).
-* New *exclude_above* parameter in *gamma2D* method. Any point in the evaluated distribution greater than exclude_above, is not accounted in the pass rate.
-* set_labeled_img method from TiffImage uses a default threshold of 90% for films identification.
+* New *clip* parameter for *to_dose* method to limit the maximum dose, (usefull for brachytherapy film dosimetry).
+* New *exclude_above* parameter for *gamma2D* method. Any point in the evaluated distribution greater than exclude_above, is not accounted in the pass rate.
+* set_labeled_img method from TiffImage uses a default threshold of 90% for films identification. Previously, it used threshold_otsu but were not useful for tif images where there is a small region of empty area relative to the film area.
 
 Issue
 ^^^^^